+++ components/ResultsPage/index.tsx
@@ -1,4 +1,4 @@
-import { Button } from '../Layout/Button'
+import { Button } from '../Forms/Button'
 import { useRouter } from 'next/router'
 import { useRef } from 'react'
 import { FieldInput } from '../../client-state/InputHelper'
